+++ packages/config/src/index.ts
@@ -1,2 +1,3 @@
 export * from './brand'
 export * from './team'
+export * from './storage'