+++ packages/config/src/storage.ts
@@ -1,0 +1,6 @@
+export class StorageConfig {
+  /**
+   * 阿里云 OSS 存储地址
+   */
+  static readonly ossUrl: string = 'https://raipiot-official-website.oss-cn-beijing.aliyuncs.com'
+}