+++ packages/config/src/team.ts
@@ -2,10 +2,10 @@ export class TeamConfig {
   /**
    * 团队名称
    */
-  static readonly name = 'raipiot Front-end Core Team'
+  static readonly name: string = 'raipiot Front-end Core Team'
 
   /**
    * 团队 Github 地址
    */
-  static readonly githubUrl = 'https://github.com/raipiot'
+  static readonly githubUrl: string = 'https://github.com/raipiot'
 }