+++ packages/antd/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @raipiot-infra/antd
 
+## 0.0.11
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @raipiot-infra/utils@0.0.7
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @raipiot-infra/utils@0.0.6
+
 ## 0.0.9
 
 ### Patch Changes