+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @raipiot-infra/config
 
+## 0.0.5
+
+### Patch Changes
+
+- update
+
+## 0.0.4
+
+### Patch Changes
+
+- update
+
 ## 0.0.3
 
 ### Patch Changes