+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @raipiot-infra/utils
 
+## 0.0.7
+
+### Patch Changes
+
+- update
+
+## 0.0.6
+
+### Patch Changes
+
+- update
+
 ## 0.0.5
 
 ### Patch Changes